fix(ci): maintainability shard — split overlay copy helpers out of auth_mounts_claude to clear 1500-line limit
auth_mounts_claude.py reached 1505 lines, 5 over the first-party
MAX_FIRST_PARTY_FILE_LINES guardrail enforced by
test_first_party_code_files_stay_under_line_limit. Extract the two
provider-neutral, symlink-safe overlay write primitives
(_safe_mtime_ns and _safe_overlay_copy) into a focused sibling module
auth_mounts_overlay_copy.py and re-import them, mirroring the existing
auth_mounts_overlay split. The functions are pure stdlib with no module
constants and are not monkeypatched on the claude module, so internal
call sites and auth_mounts_claude.<name> test references are unchanged.
